+++ Classes/Buyable.cpp
@@ -12,7 +12,7 @@ bool Buyable::get_been_bought()
     return DataManager::get_bool_from_data(this->id_key, false);
 };
 
-unsigned long long int Buyable::get_cost()
+long double Buyable::get_cost()
 {
     return this->_shop_cost;
 };
